Set dtype of pandas Series objects to float64 on initialization. This is to avoid FutureWarning.

+++ modesto/mass_flow_calculation.py
@@ -97,9 +97,9 @@ class MfCalculation(object):
 
     def initialize_mass_flows(self):
         for edge in self.edges:
-            self.mass_flows[edge] = pd.Series(0, index=self.index)
+            self.mass_flows[edge] = pd.Series(0, index=self.index, dtype='float64')
 
-        self.mass_flows[self.unknown_node][self.unknown_comp] = pd.Series(0, index=self.index)
+        self.mass_flows[self.unknown_node][self.unknown_comp] = pd.Series(0, index=self.index, dtype='float64')
 
     def calculate_mf(self):
         """
